This set of changes does three things.
Fixed the typo with http_proxy needing to be 'http_proxy'

Resolved the debug flag error, by using logging.setLevel().

It also fixes an issue I had, that may be an outlier case, where I wanted to
have the capture make an image even if the website was returning an error. This
was needed for my case because I am using this to automate screenshots of
websites I am migrating from different servers, and I want a before and after.
And sometimes, there will be passenger errors, or 503 errors, and I want those
captured.

I did change it to make a debug message instead of raise an error.