Allow loading HTML code instead of an URL
With this change you can now directly put HTML code into the library
which will then be rendered and a image will be taken of.

The methods render(), render_to_file() and render_to_bytes() do allow
either a string, which is used as URL to be fetched, like before and
as new alternative, a tuple as first parameter. The tuple needs to be
made of two strings, the first one is HTML code which is loaded into
webkit, the second one is a URL which is used as base URL for the
loaded HTML code which will be used to resolve relative URLs in the
HTML code.